+++ src/main/java/org/matsim/pt2matsim/plausibility/MappingAnalysis.java
@@ -16,7 +16,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.pt2matsim.tools;
+package org.matsim.pt2matsim.plausibility;
 
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Coord;
@@ -33,6 +33,7 @@ import org.matsim.pt.transitSchedule.api.TransitSchedule;
 import org.matsim.pt2matsim.lib.RouteShape;
 import org.matsim.pt2matsim.lib.ShapedSchedule;
 import org.matsim.pt2matsim.lib.ShapedTransitSchedule;
+import org.matsim.pt2matsim.tools.*;
 
 import java.io.FileNotFoundException;
 import java.io.UnsupportedEncodingException;
