+++ src/main/java/org/matsim/pt2matsim/tools/debug/ExtractDebugSchedule.java
@@ -16,12 +16,13 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.pt2matsim.tools;
+package org.matsim.pt2matsim.tools.debug;
 
 import org.matsim.pt.transitSchedule.api.TransitLine;
 import org.matsim.pt.transitSchedule.api.TransitRoute;
 import org.matsim.pt.transitSchedule.api.TransitRouteStop;
 import org.matsim.pt.transitSchedule.api.TransitSchedule;
+import org.matsim.pt2matsim.tools.ScheduleTools;
 
 import java.util.HashSet;
 import java.util.Set;
